requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
supabase>=1.2.0
anthropic>=0.18.1
pydantic>=2.4.2
python-dotenv>=1.0.0
